# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FitMeet",
    "description": "Backend do projeto FitMeet feito como desafio do Bootcamp da Sysmap",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/user": {
      "get": {
        "summary": "Buscar dados do usuário",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do usuário retornados com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "avatar": { "type": "string" },
                    "cpf": { "type": "string" },
                    "xp": { "type": "integer" },
                    "level": { "type": "integer" },
                    "achievements": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "criterion": { "type": "string" }
                        }
                      }
                    }
                  }
                }
              }
            },
            "401": {
              "description": "Autenticação necessaria.",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "Autenticação necessaria." }
                    }
                  }
                }
              }
            },
            "403": {
              "description": "Conta Desativada",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                    }
                  }
                }
              }
            },
            "500": {
              "description": "Erro Inesperado",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "object",
                    "properties": {
                      "error": { "type": "string", "example": "Erro Inesperado" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/preferences": {
      "get": {
        "summary": "Buscar preferencias do usuário",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Preferencias retornadas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": { "type": "string", "format": "uuid" },
                      "name": { "type": "string" },
                      "description": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/preferences/define": {
      "post": {
        "summary": "Define preferencias do usuário",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": { "type": "string", "format": "uuid" },
                "required": ["ids"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Define preferencias do usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string", "example": "Preferencias atualizadas com sucesso" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Ids invalidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Um ou mais ids informados são invalidos." }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/avatar": {
      "put": {
        "summary": "Edita foto de perfil",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["avatar"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Define preferencias do usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "avatar": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Imagem com formato invalido",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "A imagem deve ser um arquivo PNG ou JPG." }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/update": {
      "put": {
        "summary": "Edita dados de perfil",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" }
                },
                "required": ["email", "password", "name"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Define preferencias do usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "format": "uuid" },
                    "name": { "type": "string" },
                    "email": { "type": "string", "format": "email" },
                    "cpf": { "type": "string" },
                    "avatar": { "type": "string" },
                    "xp": { "type": "integer" },
                    "level": { "type": "integer" }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "409": {
            "description": "Email ou Cpf Ja pertence a outro usuario!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "O e-mail ou CPF informado já pertence a outro usuário." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/user/desactivate": {
      "delete": {
        "summary": "Desativa conta do usuario",
        "tags": ["Usuários"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Define preferencias do usuario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": "Conta desativada com sucesso"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Autenticação necessaria.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Autenticação necessaria." }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Conta Desativada",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Esta conta foi desativada e não pode ser utilizada." }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro Inesperado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": { "type": "string", "example": "Erro Inesperado" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
}